Add files via upload
Data obtained from molecular information of Caryocar brasiliense, using ISSR and SSR markers.
To examine genetic diversity, ISSR band profiles were used as dominant markers, and each was thought of as having two alleles, one that could be amplified and one that could not. The data was scored 1 for the presence and 0 for the absence of a DNA band for each loci.
For the SSR analysis, a matrix was made with rows and columns that were numbered 1, 2, and 3 to show the number of alleles for each locus. The codes 11, 22, and 33 show homozygotes, while the codes 12, 13, and 23 show heterozygotes.